requirements.txt
pyautogui
pywinauto
python-can
opencv-python
pillow
requests
pyserial
